Control up/down keys more strictly for neater animation
Before, the user could see the caret moving to the beginning/end and
then it would move to the previous line. The jittery animation didn't
look great.

If on the first line of the current editor and there is adjacent sibling
in the direction of the arrow key, prevent caret animation with return
false and move caret manually before refocusing. Same goes for
down arrow if on last line of current editor. If no adjacent sibling or
not on first/last line, act normally with return true.