Keep focus on block after un/indenting
Storing the current focused block and its caret position in redux. This
is so that when the tree changes after an un/indent and the block is
re-rendered anew (as a new DOM element), the focus will not be lost.
This may come back to bite me, but I'm not sure of another way at
the moment.