Refactor links, be conservative with setting them
Move to mirrored data structure which holds both sides of edges. This
bloats things, but allows the setting of links to be more restrictive.
Now, we can check to see if a source block's links actually changed
before going through all the logic and before checking through all of
the destination links' sources.

Only dispatch if there are links to add, if the source block is unrecorded,
and if the source block's links differ from any new links.