#7 Fix bug with Enter keypress on start of block
Pressing Enter at beginning of block correctly moved the text down to a
new sibling block. But the updateBlock check for new, incoming text to
use had a bad condition test which did not allow an empty string to be
valid block text. This meant that the current text was kept and you
ended up with duplicate lines (current line kept, new sibling with same
text).