#6 Prevent toggling unlinked refs for all pages on click
Unlinked refs were being toggled on every page after a single click on
one. This was because the <References> element in Editor.js was not
re-rendering as expected. It was sticking around and the hooks state
was shared among other pages' <References> elements. The calculation
of the unlinked references was only being done on page load, but it was
still unwelcome. Add key prop to differentiate the instances among pages.

Also fix bug where linked references were not showing because of a lapse
syncing code changes with redux.