Fix editor bugs with searching, linking; add tests
The 'searching' flag in EditBlock was fragile and didn't work as
expected. Create util function to get all of the relevant info for
determining if searching and create tests to assert behavior.

Fix bug where creating a link in the main view would break the app
if the same block was open in the sidebar.

Create test wrapper so that we can now test components with an initial
redux state.