#21 Allow arrow up/down through blocks, add tests
Fix focusedBlock bug where a caretPos of 0 was being interpreted as
false in a boolean expression

Don't set selectionStart automatically within EditBlock, when it gets
set via arrow up/down and then the user enters text, the caret goes
back to focusedBlock.caretPos and starts typing in the middle of
words.

The arrow/up down isn't high fidelity; the column does not stay the
same through the block indents, but it's good enough for now.